requirements.txt
jupyter
matplotlib
pandas
python-dateutil
scrapy
